Speed up scraper by using goroutines (#19)
* Get all repos in an organization

Currently as per docs, Repositories.List only returns
30 repos per page and on the client side we are not processing
results from the next pages. This commit adds functionality to
the client to handle paginated results from the API.

This commit also has a fix for #13 to handle a case where repo is
initialized but has no commits yet.

Closes: #13, #15

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>

* Speed up scraper using goroutines

Separate goroutines are now used for listing contributors
and commits.

Most of the time is being spent in network i/o to get this
information so the main thread is mostly sleeping. Using goroutines
helps in this case. I have noticed a 42.5% speed up on my Apple Silicon
Macbook.

Closes: #19

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>

---------

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>
Co-authored-by: Dustin Black <dblack@redhat.com>